Create palindromeNumber.js
First, we turn x (a number) to a string then have 2 pointers. One at start and one at end in which will go until the start is no longer greater than the end. Until then the points will compare the given values and if it finds any values that don't match then it will return false, else it will be true.